// src/entities/Article/model/consts/consts.ts
export enum ArticleType {
  ALL = 'ALL',
  IT = 'IT',
  DATA_SCIENCE = 'Data Science',
  DESIGN = 'Design',
  WEB_DEV = 'Web Dev',
  CYBERSECURITY = 'Cybersecurity',
  BLOCKCHAIN = 'Blockchain',
  AI = 'AI',
  MOBILE_DEV = 'Mobile Dev',
}
export enum ArticleView {
  BIG = 'BIG',
  SMALL = 'SMALL',
}

export enum ArticleSortField {
  VIEWS = 'views',
  TITLE = 'title',
  CREATED = 'created',
}

export enum ArticleBlockType {
  CODE = 'CODE',
  TEXT = 'TEXT',
  IMAGE = 'IMAGE',
}